executor: fix the unit test of the rank funcs memUsage tracking (#19896)
