+++ executor/aggfuncs/window_func_test.go
@@ -195,17 +195,3 @@ func (s *testSuite) TestWindowFunctions(c *C) {
 		s.testWindowFunc(c, test)
 	}
 }
-
-func (s *testSuite) TestMemRank(c *C) {
-	tests := []windowMemTest{
-		buildWindowMemTester(ast.WindowFuncRank, mysql.TypeLonglong, 0, 1, 1,
-			aggfuncs.DefPartialResult4RankSize, rowMemDeltaGens),
-		buildWindowMemTester(ast.WindowFuncRank, mysql.TypeLonglong, 0, 3, 0,
-			aggfuncs.DefPartialResult4RankSize, rowMemDeltaGens),
-		buildWindowMemTester(ast.WindowFuncRank, mysql.TypeLonglong, 0, 4, 1,
-			aggfuncs.DefPartialResult4RankSize, rowMemDeltaGens),
-	}
-	for _, test := range tests {
-		s.testWindowAggMemFunc(c, test)
-	}
-}
